# api/docs/swagger.yaml
definitions:
...
  genproto.MemoriesCreateReq:
    properties:
      date:
        type: string
      description:
        type: string
      location:
        type: string
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      user_id:
        type: string
    type: object
  genproto.MemoriesGetAllRes:
    properties:
      count:
        type: integer
      memories:
        items:
          $ref: '#/definitions/genproto.MemoriesRes'
        type: array
    type: object
  genproto.MemoriesGetByIdRes:
    properties:
      memory:
        $ref: '#/definitions/genproto.MemoriesRes'
    type: object
  genproto.MemoriesRes:
    properties:
      date:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        type: string
      place_name:
        type: string
      privacy:
        type: string
      tags:
        items:
          type: integer
        type: array
      title:
        type: string
      user_id:
        type: string
    type: object
  genproto.MilestonesCreateReq:
    properties:
      category:
        type: string
      date:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  genproto.MilestonesGetAllRes:
    properties:
      count:
        type: integer
      milestones:
        items:
          $ref: '#/definitions/genproto.MilestonesRes'
        type: array
    type: object
  genproto.MilestonesGetByIdRes:
    properties:
      milestone:
        $ref: '#/definitions/genproto.MilestonesRes'
    type: object
  genproto.MilestonesRes:
    properties:
      category:
        type: string
      date:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /memory:
    post:
      consumes:
      - application/json
      description: Create a new memory
      parameters:
      - description: Memory data
        in: body
        name: memory
        required: true
        schema:
          $ref: '#/definitions/genproto.MemoriesCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Memory created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create memory
      tags:
      - memory
  /memory/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a memory by ID
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Memory deleted
          schema:
            type: string
        "400":
          description: Invalid memory ID
          schema:
            type: string
        "404":
          description: Memory not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete memory
      tags:
      - memory
    get:
      consumes:
      - application/json
      description: Get a memory
      parameters:
      - description: Memory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MemoriesGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get memory
      tags:
      - memory
    put:
      consumes:
      - application/json
      description: Update an existing memory
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Title
        in: query
        name: title
        type: string
      - description: Description
        in: query
        name: description
        type: string
      - description: Privacy
        in: query
        name: privacy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Memory updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Memory not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update memory
      tags:
      - memory
...
  /memory/all:
    get:
      consumes:
      - application/json
      description: Get all memories
      parameters:
      - description: UserId
        in: query
        name: user_id
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MemoriesGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all memories
      tags:
      - memory
...
  /timeline/milestone:
    post:
      consumes:
      - application/json
      description: Create a new milestone
      parameters:
      - description: Milestone data
        in: body
        name: memory
        required: true
        schema:
          $ref: '#/definitions/genproto.MilestonesCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Milestone created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create milestone
      tags:
      - milestone
  /timeline/milestone/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a milestone by ID
      parameters:
      - description: Milestone ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Milestone deleted
          schema:
            type: string
        "400":
          description: Invalid milestone ID
          schema:
            type: string
        "404":
          description: Milestone not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete milestone
      tags:
      - milestone
    get:
      consumes:
      - application/json
      description: Get a milestone
      parameters:
      - description: Milestone ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MilestonesGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get milestone
      tags:
      - milestone
    put:
      consumes:
      - application/json
      description: Update an existing milestone
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      - description: Title
        in: query
        name: title
        type: string
      - description: Date
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Milestone updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Milestone not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update milestone
      tags:
      - milestone
  /timeline/milestone/all:
    get:
      consumes:
      - application/json
      description: Get all milestones
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.MilestonesGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all milestones
      tags:
      - milestone
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"